refactor: about component
